🚚 aswe-95 change paths and move files flask_spotify_auth.py and profile.py into src folder
