+++ src/backend/services/spotify-service/src/flask_spotify_auth.py
@@ -17,7 +17,7 @@ CLIENT_SECRET = 'b5512349cd434a04858487754e11d1e6'
 PORT = '5565'
 CALLBACK_URL = 'http://0.0.0.0'
 
-REDIRECT_URI = '{}:{}/callback/'.format(CALLBACK_URL, PORT)
+REDIRECT_URI = '{}:{}/rest/api/v1/spotify/callback/'.format(CALLBACK_URL, PORT)
 SCOPE = (
     'playlist-modify-public playlist-modify-private user-read-recently-played user-top-read '
     'user-modify-playback-state user-read-private'
